pyproject.toml
[tool.poetry]
name = "vinci-central-node"
version = "0.1.0"
description = ""
authors = ["Costa <jorgemdcosta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
pytest = "^7.4.2"
alembic = "^1.12.0"
python-decouple = "^3.8"
pydantic-settings = "^2.0.3"
protobuf = "^3.20.0"
httpx = "^0.25.0"
redis = "^5.0.0"
websockets = "^11.0.3"
pika = "^1.3.2"
python-multipart = "^0.0.6"
boto3 = "^1.28.61"
jwt = "^1.3.1"
pycryptodome = "^3.19.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
